Refactor so that we pull in less code for hello-world
