+++ packages/@ember/-internals/glimmer/lib/utils/to-bool.ts
@@ -1,7 +1,7 @@
 import { isHTMLSafe } from './string';
 import { get } from '@ember/-internals/metal/lib/property_get';
 import { tagForProperty } from '@ember/-internals/metal/lib/tags';
-import { isArray } from '@ember/array';
+import isArray from '@ember/array/lib/is-array';
 import { isProxy } from '@ember/-internals/utils/lib/is_proxy';
 import { consumeTag } from '@glimmer/validator/lib/tracking';
