+++ packages/@ember/array/index.ts
@@ -1,8 +1,6 @@
 /**
 @module @ember/array
 */
-import { DEBUG } from '@glimmer/env';
-import { PROXY_CONTENT } from '@ember/-internals/metal/lib/property_get';
 import { objectAt } from '@ember/-internals/metal/lib/object-at';
 import { replaceInNativeArray, replace } from '@ember/-internals/metal/lib/array';
 import computed from '@ember/-internals/metal/lib/computed';
@@ -17,11 +15,11 @@ import { assert } from '@ember/debug';
 import Enumerable from '@ember/enumerable';
 import MutableEnumerable from '@ember/enumerable/mutable';
 import compare from '@ember/utils/lib/compare';
-import typeOf from '@ember/utils/lib/type-of';
 import Observable from '@ember/object/observable';
 import type { MethodNamesOf, MethodParams, MethodReturns } from '@ember/-internals/utility-types';
 import type { ComputedPropertyCallback } from '@ember/-internals/metal/lib/computed';
 import { isEmberArray, setEmberArray } from '@ember/array/-internals';
+import isArray from './lib/is-array';
 
 export { default as makeArray } from './make';
 
@@ -137,65 +135,7 @@ function insertAt<T>(array: MutableArray<T>, index: number, item: T) {
   return item;
 }
 
-/**
-  Returns true if the passed object is an array or Array-like.
-
-  Objects are considered Array-like if any of the following are true:
-
-    - the object is a native Array
-    - the object has an objectAt property
-    - the object is an Object, and has a length property
-
-  Unlike `typeOf` this method returns true even if the passed object is
-  not formally an array but appears to be array-like (i.e. implements `Array`)
-
-  ```javascript
-  import { isArray } from '@ember/array';
-  import ArrayProxy from '@ember/array/proxy';
-
-  isArray();                                      // false
-  isArray([]);                                    // true
-  isArray(ArrayProxy.create({ content: [] }));    // true
-  ```
-
-  @method isArray
-  @static
-  @for @ember/array
-  @param {Object} obj The object to test
-  @return {Boolean} true if the passed object is an array or Array-like
-  @public
-*/
-export function isArray(obj: unknown): obj is ArrayLike<unknown> | EmberArray<unknown> {
-  if (DEBUG && typeof obj === 'object' && obj !== null) {
-    // SAFETY: Property read checks are safe if it's an object
-    let possibleProxyContent = (obj as any)[PROXY_CONTENT];
-    if (possibleProxyContent !== undefined) {
-      obj = possibleProxyContent;
-    }
-  }
-
-  // SAFETY: Property read checks are safe if it's an object
-  if (!obj || (obj as any).setInterval) {
-    return false;
-  }
-
-  if (Array.isArray(obj) || EmberArray.detect(obj)) {
-    return true;
-  }
-
-  let type = typeOf(obj);
-  if ('array' === type) {
-    return true;
-  }
-
-  // SAFETY: Property read checks are safe if it's an object
-  let length = (obj as any).length;
-  if (typeof length === 'number' && length === length && 'object' === type) {
-    return true;
-  }
-
-  return false;
-}
+export { isArray };
 
 /*
   This allows us to define computed properties that are not enumerable.
