+++ tests/node-vitest/tree-shakability.test.js
@@ -131,6 +131,7 @@ it('[dev] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/application/instance.js",
       "ember-source/@ember/application/namespace.js",
       "ember-source/@ember/array/index.js",
+      "ember-source/@ember/array/lib/is-array.js",
       "ember-source/@ember/array/mutable.js",
       "ember-source/@ember/array/proxy.js",
       "ember-source/@ember/canary-features/index.js",
@@ -335,6 +336,7 @@ it('[prod] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/application/instance.js",
       "ember-source/@ember/application/namespace.js",
       "ember-source/@ember/array/index.js",
+      "ember-source/@ember/array/lib/is-array.js",
       "ember-source/@ember/array/mutable.js",
       "ember-source/@ember/array/proxy.js",
       "ember-source/@ember/canary-features/index.js",
